@@ -3,20 +3,22 @@
   "author": "Eisnerw",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "description": "Software to perform generalized string interpolation",
   "license": "Apache-2.0",
   "repository": "https://github.com/eisnerw/texttemplate-editor",
   "dependencies": {
     "@webpack-cli/init": "^1.1.3",
-    "antlr4": "4.8.0",
+    "antlr4": "4.11.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "moment": "^2.29.0",
+    "moment": "^2.29.4",
     "morgan": "^1.10.0",
-    "webpack": "^4.44.2"
+    "webpack": "^5.83.1"
   },
   "devDependencies": {
+    "@types/antlr4": "^4.11.2",
     "@types/jquery": "^3.5.1",
     "@types/node": "^20.2.1",
     "file-loader": "^6.1.0",
