@@ -1,7 +1,7 @@
-const express = require('express');
+import express from 'express';
 const router = express.Router();
 
-const data = require('../data.json');
+import data from '../data.json' assert { type: 'json' };
 
 router.get('/:name', (req, res) => {
 	const { name } = req.params 
@@ -47,4 +47,4 @@ router.delete('/:id', (req, res) => {
   res.json('Successfully deleted');
 });
 
-module.exports = router;
+export {router};
