@@ -1,12 +1,11 @@
-const path = require('path');
-const express = require('express');
+import path from 'path';
+import express from 'express';
 const router = express.Router();
-const fs = require('fs');
-
+import fs from 'fs';
 
 router.get('/:name', (req, res) => {
 	const { name } = req.params 
 	res.redirect('/#' + name);
 });
 
-module.exports = router;
+export { router }
